refactor: getRandomNumber의 위치를 models 폴더로 이동
