+++ docs/README.md
@@ -77,8 +77,7 @@
 - [x] alert파일 데이터 분리 
 - [x] 주석 삭제 
 - [x] 최종 제출 코드에서 불필요한 console.log를 남기지 않는다. 
-- [ ] html에 추가한 것 지우기
-- [ ] 재시작 만들기
+- [x] html에 추가한 것 지우기
 ## 제출 전 확인하기 
 - [ ] 목록 다시 확인 
 - [ ] 피드백 다시 확인
